Tabs: activate on release + remove hover (#632)
- Remove the hover effect.
- Add a pressed effect.
- Use ButtonBase for the onClick behavior and focus state.
